Don't remove session refs on fragment runs (#9010)
It turns out #8932 was simply caused by us removing files added to the `MediaFileManager` too aggressively.
In full script runs, we can always remove session refs and clear out files from previous script runs as the whole
app is about to be redrawn anyway. When running fragments, however, doing so may leave links/buttons to a
now nonexistent file in the app, and this causes the user to hit a 404 when they click on the link/button. This PR
fixes this issue by simply not clearing file session refs during fragment runs.

Closes #8932