Fix dialogs not re-opening in some cases after dismiss (#9333)
## Describe your changes

Closes https://github.com/streamlit/streamlit/issues/9323

Add a deltaMessageReceivedAt property to our nodes to manually trigger a re-render of the Dialog component.


## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- add a JS test to make sure that the timestamp is set on the
`BlockNode`
- E2E Tests
- the existing dialog tests for reopening after dismissing should catch
any potential regression
- added a new test where a dialog is dismissed and then a different
fragment is interacted with to ensure that the dialog stays closed. This
will make sure our `deltaMessageId` stays stable

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.