Replace emails with `test.com` domain to `example.com`. See #9371 for… (#9414)
## Describe your changes

This is a follow-up PR for #9371 , where we replaced the test email
domain in just one place.

Currently, in some places, for testing purposes, we use emails with
`@test.com` domain as a placeholder. This PR replace it with
`@example.com` domain that reserverved for that purposes. For additional
context see #9371
## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

No testing needed, the changes themselves mostly in the tests, or for
test purposes.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.