Migrate Active Script hash to ScriptRunContext (#9441)
## Describe your changes

Before, the PagesManager preserves the script hash across threads which
led to race conditions with widgets resetting (See issue). The
ScriptRunContext should manage the script hashes in a thread-safe way
(because they only make sense in the context of a script run and do not
need to be preserved between script runs).

## GitHub Issue Link (if applicable)
closes #9100

## Testing Plan

- Python Unit tests are updated
- E2E test for this rapid fire implementation

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.