Re-introduce "Allow array parameters in CLI" (#9577)
## Describe your changes

**Original Description**

Streamlit has two config options that support multiple values. It works
with `config.toml`, but it does not work with the CLI arguments. This is
because we are not leveraging Click's
[multiple](https://click.palletsprojects.com/en/8.1.x/options/#multiple-options)
parameter. This change allows us to specify it. We infer the multiple
value if the default val is a list or tuple.

**Additional Changes**

I found out that we do not explicitly set defaults in click which is a
little weird. We filter out `None` values which does not affect single
values. However multiple values will get an empty tuple. We fix this by
just filtering out empty tuples because it's already either the default,
or there is no way to set it as an empty list.

Also added E2E tests. The biggest difference is that I included example
secrets and I script out the options and reset them explicitly. It's a
hack, but it makes things easier.

## Testing Plan

- Python Unit Tests updated
- E2E tests created.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.