Round pixels of calculated slider label pixel values (#9735)
## Describe your changes

In https://github.com/streamlit/streamlit/pull/9494 a dynamic way was
added to calculate the slider labels. It looks like the added snapshot
tests are flaky (on Firefox). It looks like the pixel values can have a
few decimal values which seems to shift the label slightly.

## GitHub Issue Link (if applicable)

## Testing Plan

- Fix flaky tests, so change is covered by them.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.