Add a delay before displaying the running man icon.  (#9732)
## Describe your changes

The running man animation appears while the script re-run is being
performed and for re-runs that complete quickly it introduces a flicker
and creates the perception of a longer wait. This PR introduces a delay
when showing the running man animation.

## GitHub Issue Link (if applicable)

## Testing Plan
- Unit Tests (one added to confirm not visible without delay, plus
coverage by existing running man tests)
- Manual testing by PM and Engineer to confirm expected visual
behaviour.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.