Disallow double slash paths to redirect (#9754)
## Describe your changes

We amend the call to remove the slash at the end if necessary. However,
paths with double slashes `//my-path` create an invalid redirect that
can cause security issues. This change prevents the double slash from
matching with this route.

## GitHub Issue Link (if applicable)
closes #9690

## Testing Plan

- Updated Unit Tests
- Verified manually that redirect works as needed and fails with a
double slash (and with a URL encoded double slash)

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.