Support URLPattern-compatible patterns in allowed domain matching (#9849)
## Describe your changes

Support
[URLPattern](https://developer.mozilla.org/en-US/docs/Web/API/URL_Pattern_API)-compatible
patterns in allowed domain matching.

This also fixes an issue where in Firefox domain patterns like
`https://*.streamlit.app` were never matched - this was because
previously the pattern was converted to a `URL` object by passing it to
the `URL` constructor and Firefox throws on such URLs.

The issue meant that in Firefox no messages sent from Community Cloud
apps were reaching Firefox, breaking URLs, making CC-specific menu items
not applied, etc.

In UriUtil tests, replace all `.toBeFalsy()` occurrences with
`.toBe(false)` and `.toBeTruthy()` with `.toBe(true)` to better reflect
the API is supposed to return a boolean.

Fixes #9848

## GitHub Issue Link (if applicable)

* #9848

## Testing Plan

JS Unit tests added.

I also tested it in Firefox on the app
https://app-navigation-test-app-5e8tdhuv6wvkscfe8vwxcn.streamlit.app/ by
using network overrides in Firefox DevTools and replacing the relevant
function in the minified JS code of the file
https://app-navigation-test-app-5e8tdhuv6wvkscfe8vwxcn.streamlit.app/~/+/static/js/main.75ac1cb6.js.
After the refresh, the problems described in issue #9848 disappeared.

**NOTE:** There's one change - previously, the pattern `"https://*.com"`
would not match `https://example.example.com` and now it does. If one
wants to forbid it, a more detailed regex pattern can be used, e.g.
`"https://(\w+).com"`.

As I understand, for now the list of allowed domains is not exposed as a
public API, so perhaps such a change is fine.

There's an open issue for this
(https://github.com/streamlit/streamlit/issues/6389) and a PR for it
labded (https://github.com/streamlit/streamlit/pull/7180) but to the
`feature/host-security-config` branch.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.