Remove temporary padding fix for fullscreen button (#9882)
## Describe your changes

Removes a temporary fix that was added to prevent horizontal scrolling
because of the fullscreen action button. Since we have replaced all
fullscreen action buttons with the toolbar, the temporary fix is no more
necessary.

It also addresses an issue where the wide mode can have more padding
than a centered mode on small screen sizes.

## GitHub Issue Link (if applicable)

- Closes: https://github.com/streamlit/streamlit/issues/9707

## Testing Plan

- Updated snapshots and tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.