Remove internal Streamlit details from exception traces  (#9913)
## Describe your changes

When the user's code causes an exception to be raised, today we show a
bunch of internal Streamlit runtime-related code in the traceback which
doesn't help the user in any way. In fact, if just makes debugging
harder.

For example, here's a script that throws `RuntimeError("OMG")`:

![Screenshot 2024-11-22 at 17 42
02](https://github.com/user-attachments/assets/a6d5ba70-e16d-445d-820b-699692dae4b7)

But if we remove all those runtime lines, suddenly everything is a whole
lot easier to debug:
 
![Screenshot 2024-11-22 at 17 42
27](https://github.com/user-attachments/assets/9d14a735-9f33-4fed-a1b0-cd02b97f6acd)

Keep in mind, this is just a toy example. In reality, tracebacks are
more complex. But this change makes them much simpler to understand!

What this PR does:

1. Anytime we're drawing an `st.exception` we take that traceback and
split it into two: the "internal" frames, which don't include any
user-code, and the "external" frames, which start from a user-code frame
and include everything after that.

1. Removes some old logic where we tried to simplify the traceback when
the exception was a `StreamlitAPIError`. But the logic from (1) subsumes
that.

1. Adds a `.main_script_parent` property to `PagesManager` and
`ScriptContext`, since we need that pretty often. (And I needed it for
(1) too)

## GitHub Issue Link (if applicable)

## Testing Plan

- ~~Explanation of why no additional tests are needed~~
- **Unit Tests (JS and/or Python):** ✅ Included!
- **E2E Tests:** New ones not needed. Will update screenshots, though.
- **Any manual testing needed?** Not needed

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.