Do not include source map files in release builds (#10011)
## Describe your changes

The move to Vite accidentally included source map files in the release
build. This bloats the package size by a large amount. This change
modifies `BUILD_AS_FAST_AS_POSSIBLE` to be `DEV_BUILD` where users can
decide whether to enable source maps for the production build with
`DEV_BUILD=1 make frontend`. As a result `make frontend` and `make
frontend-fast` are the same, so we remove the `frontend-fast` option.

Note `BUILD_AS_FAST_AS_POSSIBLE` is the opposite of `DEV_BUILD`, so we
update the sourcemap config to be whether we want a DEV_BUILD or not.

## GitHub Issue Link (if applicable)
closes #10008

## Testing Plan

- Tests should pass (and manually checked file size of wheel build) as
sourcemaps don't affect the build.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.