Add back `CONTRIBUTING.md` with shortened text (#10058)
## Describe your changes

https://github.com/streamlit/streamlit/pull/10051 added info about
contributing to Streamlit to the Readme and removed the
`CONTRIBUTING.md` file. However, the `CONTRIBUTING.md` is one of the
[default repo health
files](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file)
in Github with some deeper integrations. Therefore, it might still be
useful to keep this file in our repo. This PR adds it back with a
shortend version that links to our wiki.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Debbie Matthews <debbie.matthews@snowflake.com>