Enforce d3-color 3.1 dependency (#10142)
## Describe your changes

A previous attempt to update the d3-color dependency to 3.1 in
https://github.com/streamlit/streamlit/pull/10109 only resolved one of
the issues (with d3-graphviz). Baseweb UI still relies on `d3` v6.7,
which requires `d3-color` v2. However, the `d3` component only seems to
be used as part of the data-table -> which is unused in Streamlit.
Therefore, it's pretty safe for us to enforce at least d3-color 3.1 via
`resolutions`.

https://github.com/streamlit/streamlit/security/dependabot/91

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.