split up WebsocketConnection.tsx to pass lint rule (#10177)
## Describe your changes

In preparation for adding the [react-refresh
rule](https://github.com/ArnaudBarre/eslint-plugin-react-refresh), this
PR splits up WebsocketConnection.tsx .

Further PRs are in the works for fixing other violations of this linter
rule and then adding the linter rule.

Note, this file should not break fast refresh, so the linting rule is a
false positive here, but I thought it could still be good to take the
opportunity to reduce the file size a bit.

## GitHub Issue Link (if applicable)

## Testing Plan

Does not change any functionality so this is covered by existing tests. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.