Preserve scroll position for `st.selectbox` and `st.multiselect` (#10073)
## Describe your changes

When you click on a selectbox (or multiselect), scroll in the dropdown,
then close it and click on the selectbox again, the dropdown opens at
the top position. This can be annoying if the dropdown is large. With
this PR, the dropdown will open at the same position where it was
previously closed. Also, when the dropdown opens for the first time, it
opens at the position of the selected default value.

**Before:**


https://github.com/user-attachments/assets/7b361fb1-0e0d-47db-b627-5391455a2228


**After:**


https://github.com/user-attachments/assets/69ca577a-d78e-43fb-8a26-11ec56003638


## GitHub Issue Link (if applicable)

Closes #4901

## Testing Plan

- Added a JS unit test on the `Selectbox` component that simulates
scrolling, closes the dropdown, and checks that it opens at the same
position.
- Did not add an e2e test because it would basically do the same thing
and that feels a bit overkill. But I can add e2e test for `st.selectbox`
and `st.multiselect` if we want to have them.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Johannes Rieke <johannes.rieke@snowflake.com>